+++ packages/zeebe-element-templates-json-schema/test/spec/validationSpec.js
@@ -525,6 +525,8 @@ describe('validation', function() {
 
       it('called-element-applies-to-only');
 
+      it('called-element-business-id');
+
       it('called-element-invalid-applies-to-no-element-type');
 
       it('called-element-invalid-element-type');