+++ packages/zeebe-element-templates-json-schema/CHANGELOG.md
@@ -6,6 +6,8 @@ All notable changes to [@camunda/zeebe-element-templates-json-schema](https://gi
 
 ___Note:__ Yet to be released changes appear here._
 
+* `FEAT`: allow `businessId` as `zeebe:calledElement` binding property ([#5912](https://github.com/camunda/camunda-modeler/issues/5912))
+
 ## 0.42.0
 
 * `FEAT`: support `isEmpty` condition ([#252](https://github.com/camunda/element-templates-json-schema/pull/252))